expol: merge files; initially reviewed the src dir
