+++ src/hippo/expol.cpp
@@ -1,6 +1,5 @@
 #include "ff/hippo/expol.h"
 #include "ff/atom.h"
-#include "ff/energybuffer.h"
 #include "ff/hippomod.h"
 #include "tool/darray.h"
 #include "tool/iofortstr.h"
@@ -49,8 +48,8 @@ void alterpol(real (*polscale)[3][3], real (*polinv)[3][3])
    TINKER_FCALL2(acc1, cu1, alterpol, polscale, polinv);
 }
 
-TINKER_FVOID2(
-   acc1, cu1, dexpol, int, const real (*)[3], grad_prec*, grad_prec*, grad_prec*, VirialBuffer);
+TINKER_FVOID2(acc1, cu1, dexpol, //
+   int, const real (*)[3], grad_prec*, grad_prec*, grad_prec*, VirialBuffer);
 void dexpol(int vers, const real (*uind)[3], grad_prec* depx, grad_prec* depy, grad_prec* depz,
    VirialBuffer vir_ep)
 {
